backend/requirements.txt
# Web framework
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0

# Google Auth
google-auth==2.29.0

# Data processing
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2        # Excel support
pyarrow==16.0.0         # Parquet support

# Machine Learning
scikit-learn==1.4.2
joblib==1.4.0
statsmodels==0.14.2     # ARIMA

# Visualization
plotly==5.21.0

# NLP (optional — comment out if not needed)
transformers==4.40.2
torch==2.3.0            # CPU-only: pip install torch --index-url https://download.pytorch.org/whl/cpu

# PDF export
reportlab==4.2.2

# Utilities
python-dotenv==1.0.1
werkzeug==3.0.3
